+++ backend/core/src/auth/controllers/auth.controller.ts
@@ -3,9 +3,10 @@ import { LocalAuthGuard } from "../guards/local-auth.guard"
 import { AuthService } from "../services/auth.service"
 import { DefaultAuthGuard } from "../guards/default.guard"
 import { ApiBody, ApiOperation, ApiTags } from "@nestjs/swagger"
-import { LoginDto, LoginResponseDto } from "../dto/login.dto"
+import { LoginDto } from "../dto/login.dto"
 import { mapTo } from "../../shared/mapTo"
 import { defaultValidationPipeOptions } from "../../shared/default-validation-pipe-options"
+import { LoginResponseDto } from "../dto/login-response.dto"
 
 @Controller("auth")
 @ApiTags("auth")