+++ backend/core/src/auth/dto/user-roles-create.dto.ts
@@ -1,0 +1,4 @@
+import { OmitType } from "@nestjs/swagger"
+import { UserRolesDto } from "./user-roles.dto"
+
+export class UserRolesCreateDto extends OmitType(UserRolesDto, ["user"] as const) {}